requirements.txt
fastapi==0.109.2
alice==0.0.1
alice_scripts==0.2.post1
annotated-types==0.6.0
anyio==4.2.0
beautifulsoup4==4.12.3
blinker==1.7.0
bs4==0.0.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
DAWG-Python==0.7.2
docopt==0.6.2
exceptiongroup==1.2.0
fastapi==0.109.2
Flask==3.0.2
Flask-Cors==4.0.0
h11==0.14.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
lxml==5.1.0
MarkupSafe==2.1.5
pydantic==2.6.1
pydantic_core==2.16.2
pymorphy2==0.9.1
pymorphy2-dicts-ru==2.4.417127.4579844
python-docx==1.1.0
python-multipart==0.0.9
requests==2.31.0
sniffio==1.3.0
soupsieve==2.5
starlette==0.36.3
typing_extensions==4.9.0
urllib3==2.2.1
uvicorn==0.27.1
Werkzeug==3.0.1